+++ UiPathOrch/OrchestratorCmdlet/ApiTriggerCmdlet/CopyApiTrigger.cs
@@ -33,6 +33,7 @@ public class CopyApiTriggerCmdlet : OrchestratorPSCmdlet
         var srcDrivesFolders = SessionState.EnumFolders(Path, Recurse.IsPresent, Depth);
 
         var (dstDrive, dstRootFolder) = SessionState.ResolveToSingleFolder(Destination);
+        var dstFolderCache = new Dictionary<string, Folder?>();
 
         // If source and destination are the same, do nothing
         if (srcRootFolder == dstRootFolder) return;
@@ -56,7 +57,7 @@ public class CopyApiTriggerCmdlet : OrchestratorPSCmdlet
                 continue;
             }
 
-            Folder? dstFolder = this.GetRelativeDstFolder(srcRootFolder, srcFolder, dstDrive, dstRootFolder);
+            Folder? dstFolder = this.GetRelativeDstFolder(srcRootFolder, srcFolder, dstDrive, dstRootFolder, createIfMissing: true, createCache: dstFolderCache);
             if (dstFolder is null || srcFolder == dstFolder) continue;
 
             //srcDrive._dicReleases?.TryRemove(srcFolder.Id ?? 0, out _);