+++ UiPathOrch/OrchestratorCmdlet/AssetCmdlet/CopyAsset.cs
@@ -41,6 +41,7 @@ public class CopyAssetCmdlet : OrchestratorPSCmdlet
         var srcDrivesFolders = SessionState.EnumFolders(Path, Recurse.IsPresent, Depth);
 
         var (dstDrive, dstRootFolder) = SessionState.ResolveToSingleFolder(Destination);
+        var dstFolderCache = new Dictionary<string, Folder?>();
 
         // If the source and destination are the same, do nothing
         if (srcRootFolder == dstRootFolder) return;
@@ -74,7 +75,7 @@ public class CopyAssetCmdlet : OrchestratorPSCmdlet
                 continue;
             }
 
-            Folder? dstFolder = this.GetRelativeDstFolder(srcRootFolder, srcFolder, dstDrive, dstRootFolder);
+            Folder? dstFolder = this.GetRelativeDstFolder(srcRootFolder, srcFolder, dstDrive, dstRootFolder, createIfMissing: true, createCache: dstFolderCache);
             if (dstFolder is null || srcFolder == dstFolder) continue;
 
             try