+++ UiPathOrch/OrchestratorCmdlet/BucketCmdlet/CopyBucket.cs
@@ -34,6 +34,7 @@ public class CopyBucketCmdlet : OrchestratorPSCmdlet
         var srcDrivesFolders = SessionState.EnumFolders(Path, Recurse.IsPresent, Depth);
 
         var (dstDrive, dstRootFolder) = SessionState.ResolveToSingleFolder(Destination);
+        var dstFolderCache = new Dictionary<string, Folder?>();
 
         // If source and destination are the same, do nothing
         if (srcRootFolder == dstRootFolder) return;
@@ -58,7 +59,7 @@ public class CopyBucketCmdlet : OrchestratorPSCmdlet
                 continue;
             }
 
-            Folder? dstFolder = this.GetRelativeDstFolder(srcRootFolder, srcFolder, dstDrive, dstRootFolder);
+            Folder? dstFolder = this.GetRelativeDstFolder(srcRootFolder, srcFolder, dstDrive, dstRootFolder, createIfMissing: true, createCache: dstFolderCache);
             if (dstFolder is null || srcFolder == dstFolder) continue;
 
             try