+++ UiPathOrch/OrchestratorCmdlet/TestDataQueueCmdlet/CopyTestDataQueue.cs
@@ -33,6 +33,7 @@ public class CopyTestDataQueueCmdlet : OrchestratorPSCmdlet
         var srcDrivesFolders = SessionState.EnumFoldersWithoutPersonalWorkspace(Path, Recurse.IsPresent, Depth);
 
         var (dstDrive, dstRootFolder) = SessionState.ResolveToSingleFolder(Destination);
+        var dstFolderCache = new Dictionary<string, Folder?>();
 
         // If source and destination are the same, do nothing
         if (srcRootFolder == dstRootFolder) return;
@@ -57,7 +58,7 @@ public class CopyTestDataQueueCmdlet : OrchestratorPSCmdlet
                 continue;
             }
 
-            Folder? dstFolder = this.GetRelativeDstFolder(srcRootFolder, srcFolder, dstDrive, dstRootFolder);
+            Folder? dstFolder = this.GetRelativeDstFolder(srcRootFolder, srcFolder, dstDrive, dstRootFolder, createIfMissing: true, createCache: dstFolderCache);
             if (dstFolder is null || srcFolder == dstFolder) continue;
 
             try